pyproject.toml
[project]
name = "dr"
version = "0.1.0"
description = "Google Drive file copy utilities"
requires-python = ">=3.8"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "deepface>=0.0.79",
    "tensorflow>=2.13.0",
    "tf-keras>=2.13.0",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
    "faiss-cpu>=1.7.4",
    "pillow>=10.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["face_extraction.py"]

[tool.uv]
dev-dependencies = []
